requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.0
opencv-python==4.10.0.84
pillow==11.0.0
pyaudio==0.2.14
pygments==2.18.0
rich==13.9.4
six==1.17.0
websockets==14.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
aiosqlite==0.19.0